refactor: :construction: Move common events handling into a separate function
